refactor: split linter presentation from CLI Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
